requirements.txt
alpha-vantage==2.3.1
APScheduler==3.8.1
asgiref==3.4.1
asyncio==3.4.3
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
channels==3.0.3
charset-normalizer==2.0.6
coverage==6.0
daphne==3.0.2
Django==3.2.7
django-cors-headers==3.9.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
